made ppu stuff slightly more accurate in a futile attempt to fix smb1's hud
it actually does sprite evaluation stuff now and can run a lot faster, probably should've been doing that much sooner lmao

+++ src/cpu.h
@@ -25,7 +25,7 @@ typedef struct {
 	uint8_t s;
 	uint8_t p;
 	uint8_t irq;
-	uint16_t cycles;
+	uint64_t cycles;
 } cpu_t;
 
 extern cpu_t cpu;
